+++ controllers/examples/scene_initialization.py
@@ -6,7 +6,7 @@ This is a simple example of how to initialize an interior environment populated 
 
 if __name__ == "__main__":
     # Instantiate the controller.
-    m = Magnebot(launch_build=False)
+    m = Magnebot()
     # Initialize the scene. Populate it with objects. Spawn the Magnebot in a room.
     m.init_scene(scene="2a", layout=1, room=1)
     # Show the rendered image.